requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
pymongo~=4.15.0
python-dotenv~=1.1.1
pydantic~=2.11.9